Use when-let in `projectile-with-default-dir'
Many projectile commands pass the result of `projectile-ensure-project' directly
to this macro. The problem is that `projectile-ensure-project' can still yield
nil if the user cancels the project selection dialog. This causes all such
functions to error with "error: stringp nil" if the user has to select a project
and cancels.